Leaves the current implementation as one that works for fixed-length problems
