+++ src/poli_baselines/solvers/__init__.py
@@ -1,6 +1,6 @@
 from .simple.random_mutation import RandomMutation
 from .simple.continuous_random_mutation import ContinuousRandomMutation
-from .simple.genetic_algorithm import GeneticAlgorithm
+from .simple.genetic_algorithm import FixedLengthGeneticAlgorithm
 
 from .bayesian_optimization.vanilla_bayesian_optimization import (
     VanillaBayesianOptimization,
