+++ src/poli_baselines/tests/solvers/simple/genetic_algorithm/test_genetic_algorithm.py
@@ -8,29 +8,32 @@ import pytest
 
 def test_genetic_algorithm_interface():
     from poli.objective_repository import AlohaProblemFactory
-    from poli_baselines.solvers import GeneticAlgorithm
+    from poli_baselines.solvers import FixedLengthGeneticAlgorithm
 
     f, x0, y0 = AlohaProblemFactory().create()
 
-    solver = GeneticAlgorithm(black_box=f, x0=x0, y0=y0, pop_size=10)
+    solver = FixedLengthGeneticAlgorithm(black_box=f, x0=x0, y0=y0, pop_size=10)
 
     solver.solve(max_iter=100)
 
 
 def test_genetic_algorithm_improves_over_time():
     from poli.objective_repository import AlohaProblemFactory
-    from poli_baselines.solvers import GeneticAlgorithm
+    from poli_baselines.solvers import FixedLengthGeneticAlgorithm
 
     f, x0, y0 = AlohaProblemFactory().create()
 
-    solver = GeneticAlgorithm(black_box=f, x0=x0, y0=y0, pop_size=10)
+    solver = FixedLengthGeneticAlgorithm(black_box=f, x0=x0, y0=y0, pop_size=10)
 
-    solver.solve(max_iter=10)
+    solver.solve(max_iter=20)
     assert y0 < solver.get_best_performance()
 
 
 def test_genetic_algorithm_on_foldx_stability_with_several_wildtypes():
-    from poli_baselines.solvers import GeneticAlgorithm
+    from poli_baselines.solvers import FixedLengthGeneticAlgorithm
+    from poli.core.util.seeding import seed_python_numpy_and_torch
+
+    seed_python_numpy_and_torch(13)
 
     register = pytest.importorskip("poli.objective_repository.foldx_stability.register")
 
@@ -40,11 +43,11 @@ def test_genetic_algorithm_on_foldx_stability_with_several_wildtypes():
 
     problem_factory = register.FoldXStabilityProblemFactory()
     f, x0, y0 = problem_factory.create(
-        wildtype_pdb_path=wildtype_pdb_paths, verbose=True
+        wildtype_pdb_path=wildtype_pdb_paths, verbose=False
     )
 
-    solver = GeneticAlgorithm(black_box=f, x0=x0, y0=y0, pop_size=5)
-    solver.solve(max_iter=2)
+    solver = FixedLengthGeneticAlgorithm(black_box=f, x0=x0, y0=y0, pop_size=4)
+    solver.solve(max_iter=2, verbose=True)
     print(solver.get_best_solution())
